kernel-3.2: move atomic_flags field to bottom of the task_struct structure
Adding a field of datatype 'unsigned long' in the middle between the other 'tasks states' is too big
and causes a kernel panic.
Field 'atomic_flags' may have a datatype of 'short' but not int, long or long long.
Or a bit field with a depth of 16 bit or fewer.
This is probably caused by added code by the 001-linux-dreambox-kernel.patch.

But the arguments of functions set_bit(),clear_bits() and test_bits() expecting a datatype of 'unsigned long'.
Well, simply moving atomic_flags to the buttom of structure task_struct
causing no kernel panic and the box boot flawlessly.

Remove the 2 patches which reverts commits:
https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-stable.git/commit/?h=v3.2.101&id=4fce5e64991fc0a02e47e525ce72f33def17b7ef
and
https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-stable.git/commit/?h=v3.2.101&id=4fce5e64991fc0a02e47e525ce72f33def17b7ef