Apply framebuffer patch
Based on:
https://github.com/openatv/enigma2/commit/0f049825127cc7df976f3c4a7787894f64f6a6ef

Without macro and preprocessor directives since we apply this patch for given machines only, is that a bit ambiguous.
Also to keep the patch as small as possible to prevent patch conflicts.

This addresses the problem with non full hd skins (skin disappear/reappear after reboot).
Thanks Zeros for testing.