!PLEASE READ!
The secondstage recipes for dm800, dm800se and dm500hd are now dummy packages!
I removed the dependency on mtd-utils to free up space.
This means that reinstalling or updating it won't reflash the box any longer.
This is OK as these receivers are out of support for quite some time.
DMM won't release new bootloaders for these receivers anyways.
The packages are now only used to stage the secondstage binary into the DEPLOYDIR for image creation.